refactor: :truck: move avatax-order-created- to its own folder
